pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "CnnWasteDetector"
version = "0.1.0"
description = "Project aimed to develop neural net capable of identyfing different types of trash"
authors = ["Mikołaj Czachorowski"]

[tool.poetry.dependencies]
python = "^3.9"
pillow = ">=10.4.0"
jupyterlab = ">=3.6"
pandas = ">=2.0"
matplotlib = ">=3.7"
opencv-python = ">=4.10.0.84"
fire = ">=0.6.0 "
ruff = ">=0.1"
pre-commit = ">=2.20"
torchvision = "0.19.0"

[tool.poetry.extras]
jupyter = ["jupyterlab"]
lint = ["ruff", "pre-commit"]


[tool.poetry.urls]
homepage = "https://github.com/Micz26/CNN-Waste-Detector.git"
repository = "https://github.com/Micz26/CNN-Waste-Detector.git"


[tool.ruff]
select = ["E", "W", "F", "B", "C", "I"]
line-length = 88

ignore = [
    # "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]
extend-exclude = [".venv/", "build/", "dist/", "docs/"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint.isort]
order-by-type = true
lines-between-types = 1
combine-as-imports = true
# known_first_party = "common"
